chore: cleanup fast-element interfaces
